+++ STRATEGY/src/pckg_VJ_Z1_a/Kalkulacija.java
@@ -1,5 +1,5 @@
 package pckg_VJ_Z1_a;
 
 public interface Kalkulacija {
-    Double performCalculation(Double fst, Double snd);
+    float performCalculation(float fst, float snd);
 }
